% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{abort_packages_not_installed}
\alias{abort_packages_not_installed}
\title{Throw an error if required packages are not installed.}
\usage{
abort_packages_not_installed(...)
}
\description{
Reports which packages need to be installed and the parent function name.
See \url{https://stackoverflow.com/questions/15595478/how-to-get-the-name-of-the-calling-function-inside-the-called-routine}
This is only intended to be used inside a function. It will error otherwise.
}
\details{
Adapted from \href{abort_packages_not_installed}{mikropml}
}
\examples{
abort_packages_not_installed("base")
\dontrun{
abort_packages_not_installed("not-a-package-name", "caret", "dplyr", "non_package")
}
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
