% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{get_deps_tbl}
\alias{get_deps_tbl}
\title{Get package dependencies as a tibble}
\usage{
get_deps_tbl(package)
}
\arguments{
\item{package}{name of an R package}
}
\value{
a tibble with the package's recursive dependencies
}
\description{
Get package dependencies as a tibble
}
\examples{
get_deps_tbl('mikropml')
}
