% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{get_height_inches}
\alias{get_height_inches}
\title{Get image height in inches given dimensions in pixels and width in inches}
\usage{
get_height_inches(width_pixels, height_pixels, width_inches = 6.5)
}
\description{
Get image height in inches given dimensions in pixels and width in inches
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
