% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{get_image_dims}
\alias{get_image_dims}
\title{Get image dimensions in inches from a file path}
\usage{
get_image_dims(image_path, width_inches = 6.5)
}
\description{
Get image dimensions in inches from a file path
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
