% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{insert_image_md}
\alias{insert_image_md}
\title{Insert image in markdown}
\usage{
insert_image_md(image_path, width_inches = 6.5)
}
\description{
Insert image in markdown
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
