% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sovatools-package.R
\docType{package}
\name{sovatools-package}
\alias{sovatools}
\alias{sovatools-package}
\title{sovatools: Sovacool's Tools}
\description{
A collection of useful functions and examples that don't have an obvious home anywhere else. See the website \url{https://sovacool.dev/sovatools/} for more information, documentation, and examples.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kelly-sovacool/sovatools/}
  \item \url{https://sovacool.dev/sovatools/}
  \item Report bugs at \url{https://github.com/kelly-sovacool/sovatools/issues}
}

}
\author{
\strong{Maintainer}: Kelly Sovacool \email{sovacool@umich.edu} (\href{https://orcid.org/0000-0003-3283-829X}{ORCID})

}
\keyword{internal}
