% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sovatools.R
\docType{package}
\name{sovatools}
\alias{sovatools}
\title{sovatools: Sovacool's Tools}
\description{
A collection of useful functions and example code that doesn't
have an obvious home anywhere else. See the
\href{http://sovacool.dev/sovatools}{website}
for more information, documentation, and examples.
}
