% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{strip_leading_slash}
\alias{strip_leading_slash}
\title{Strip the leading slash from a string}
\usage{
strip_leading_slash(x)
}
\description{
Strip the leading slash from a string
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
