% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_sovatools}
\alias{theme_sovatools}
\title{sovatools theme for ggplot}
\usage{
theme_sovatools()
}
\value{
list of ggproto objects
}
\description{
Uses \code{ggplot2::theme_bw()} and removes margins.
}
\examples{
library(ggplot2)
ggplot(mtcars) +
  aes(x = mpg, y = wt, color = cyl) +
  geom_point()  +
  theme_sovatools()
}
